+++ crates/bcr-ebill-wasm/src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::arc_with_non_send_sync)]
 use bcr_ebill_api::{Config as ApiConfig, get_db_context, init};
+use constants::SURREAL_DB_CON_INDXDB_DATA;
 use context::{Context, get_ctx};
 use futures::{StreamExt, future::ready};
 use gloo_timers::future::{IntervalStream, TimeoutFuture};
@@ -24,8 +25,6 @@ pub struct Config {
     pub log_level: Option<log::Level>,
     pub bitcoin_network: String,
     pub nostr_relay: String,
-    pub surreal_db_connection: String,
-    pub data_dir: String,
     pub job_runner_initial_delay_seconds: u32,
     pub job_runner_check_interval_seconds: u32,
 }
@@ -33,7 +32,7 @@ pub struct Config {
 pub type Result<T> = std::result::Result<T, error::WasmError>;
 
 thread_local! {
-    static CONTEXT: RefCell<Option<Context>> = const { RefCell::new(None) } ;
+    static CONTEXT: RefCell<Option<&'static Context>> = const { RefCell::new(None) } ;
 }
 
 #[wasm_bindgen]
@@ -50,8 +49,8 @@ pub async fn initialize_api(
     let api_config = ApiConfig {
         bitcoin_network: config.bitcoin_network,
         nostr_relay: config.nostr_relay,
-        surreal_db_connection: config.surreal_db_connection,
-        data_dir: config.data_dir,
+        surreal_db_connection: SURREAL_DB_CON_INDXDB_DATA.to_owned(),
+        data_dir: "./".to_owned(), // unused in wasm
     };
     init(api_config.clone())?;
 
@@ -64,12 +63,13 @@ pub async fn initialize_api(
     info!("Local npub as hex: {:?}", keys.get_nostr_npub_as_hex());
     info!("Config: {api_config:?}");
 
-    // init context
-    let ctx = Context::new(api_config.clone(), db, &keys.get_public_key()).await?;
+    // init context as static reference
+    let ctx = Context::new(api_config.clone(), db).await?;
     CONTEXT.with(|context| {
         let mut context_ref = context.borrow_mut();
         if context_ref.is_none() {
-            *context_ref = Some(ctx);
+            let leaked: &'static Context = Box::leak(Box::new(ctx)); // leak to get a static ref
+            *context_ref = Some(leaked);
         }
     });